// src/types/interface/health.ts
//定义所需人员信息的属性
export interface PersonCard {
    //看接口需求
}

//定义人员属性的类型
export interface PersonCardType {
    bigGanglion?: string;
    brucella?: string;
    child?: string;
    coronary?: string;
    deformed?: string;
    dentalFluorosis?: string;
    echinococcosis?: string;
    familyDoctor?: string;
    generalPopulation?: string;
    hypertension?: string;
    iodine?: string;
    maternal?: string;
    mentalDisorder?: string;
    [key: string]: any;
}